feat: treat returned response data dynamically
BREAKING CHANGE: the responseType option has been removed.
to return a raw response, return an object, containing the properties raw: true and data: <data here>
to not send a response either return an object with data: undefined or nothing at all

BREAKING CHANGE: redirects have been reworked.
to redirect, return an object with the redirect property set to the path you want to redirect to

BREAKING CHANGE: responses are now treated differently depending on what data is returned.
if you have methods returning objects with a status, a redirect or a data field, you need to migrate them
by encapsulating them in another object under the data key, so they are treated as data and not response properties.